+++ hotel-booking-frontend/src/main.tsx
@@ -2,9 +2,11 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
+import "vanilla-cookieconsent/dist/cookieconsent.css";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { AppContextProvider } from "./contexts/AppContext.tsx";
 import { SearchContextProvider } from "./contexts/SearchContext.tsx";
+import { initCookieConsent } from "./lib/cookie-consent.ts";
 
 export const queryClient = new QueryClient({
   defaultOptions: {
@@ -13,6 +15,9 @@ export const queryClient = new QueryClient({
     },
   },
 });
+
+initCookieConsent();
+
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>